requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-ninja==1.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
